requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
deepface
tensorflow
supabase
numpy
Pillow
python-dotenv
gunicorn
